Add files via upload
This code will allow the user to enter their full name with "First, Middle, Last" as the format. First, it will capitalize the starting letter of each name, then set the middle initial only with a period after it, lastly, it will change the format of the name into "Last name, First name M." and display it. So with the given problem which is "juan,carlos,dela cruz" the displayed name will now be "Dela Cruz, Juan C."